requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==4.0.1
django-cors-headers==3.11.0
django-extensions==3.1.5
djangorestframework==3.13.1
gunicorn==20.1.0
mysqlclient==2.1.0
pydot==1.4.2
pyparsing==3.0.7
python-decouple==3.6
pytz==2021.3
sqlparse==0.4.2
fpdf2==2.5.5
django-heroku==0.3.1
requests==2.28.1
django-rest-knox==4.2.0